The Tale of the Terrifying Tangents: There was once an Engine called Topaz. Topaz had children. Their names were Positions, Texcoords, Normals, Tangents and Indices. One day, Topaz had realised that for the entirety of her life, Indices had not been defined correctly; infact being partially defined relative to Tangents. Topaz, and his creator, Harrand, mystified that Indices never appeared malformed despite such a huge flaw, corrected this issue. Despite, programmatically, Indices was affected by Tangents, no noticeable difference in Indices affected Tangents. Topaz was forced to murder Tangents. Indices, shockingly, experienced no change and worked as before. In an even more terrifyingly inexplicable event, Tangents' school still sent Topaz pictures that Tangents had drawn. How could a feature that, not only was invalid for years and then REMOVED continue to produce the effect that it would have produced had it had been developed correctly?
In actual English:
Removed all implementation of tangents. No tangent data is being sent to the GPU VRAM. However, the implementation of tangents in the shaders remains. It turns out, that my Parallax Mapping never worked properly at all. Upon removal of tangents, DESPITE THE SHADER NOW USING DATA THAT NO LONGER FUCKING EXISTS, now works correctly and looks beautiful.
The attribPointer index that tangents was using was 3. Upon removal, I gave that index to indices. THAT BROKE IT. 3 now remains unused as an index and tangents seem to magically appear out of nowhere. Praise the OpenGL Gods, PRAISE THEM (Please help me I thought I understood this, it's like Paranormal Activity 0x06 over here).
And that is the Tale of the Terrifying Tangents. I thought perhaps the tangent data stayed in VRAM from a previous test build but I restarted and the thing still works. PRAISE THEM, I WON'T TOUCH MY NOW-INVALID-YET-FUNCTIONAL-SHADERS EVER AGAIN. PRAISE THEM.
In other news, I think I've completely lost it. Thanks for reading this essay of a commit message; it was necessary because this is spooky as hell.

Edit: FPS ACROSS THE BOARD HAS INCREASED BECAUSE OF THIS WHAT THE FUCK?