Improved shadow mapping so that the projection matrix generated by a light can always pack to fit a scene perfectly. Also added tz::graphics::BufferBit as a type-safe container for the enums like GL_DEPTH_BUFFER_BIT (and graphics API independent, if support for Vulkan or D3D is ever implemented).
One of two bugs with shadows have been fixed. The one where objects outside of the shadowmap are always in shadow has been fixed.
However, the bug where increases in the positive z or positive y axes in object positions cause them not to cast shadows despite being in range for the shadow map. This is very odd, and I suspect my implementation is wrong in some way (perhaps pre-multiply view and projection matrices for a lightspace matrix BEFORE the depth shader does it?)